+++ mongo_datatables/__init__.py
@@ -1,9 +1,7 @@
-"""
-MongoDB integration with jQuery DataTables
-"""
+"""Integrate MongoDB with jQuery DataTables server-side processing."""
 __version__ = '2.0.0'
 
-from mongo_datatables.datatables import DataTables, DataField
+from mongo_datatables.datatables_core import DataTables, DataField
 from mongo_datatables.editor import Editor, StorageAdapter
 from mongo_datatables.exceptions import (
     MongoDataTablesError,